fix: eventual consistency for sync creation
