+++ apps/api/services/index.ts
@@ -1,4 +1,3 @@
 export * from './application_service';
-export * from './connection_writer_service';
+export * from './connection_and_sync_service';
 export * from './passthrough_service';
-export * from './sync_service';
