+++ openapi/mgmt/components/schemas/create_update_integration.yaml
@@ -14,3 +14,4 @@ required:
   - category
   - auth_type
   - provider_name
+  - config
