+++ packages/core/index.ts
@@ -3,3 +3,4 @@ export * as lib from './lib';
 export * as mappers from './mappers';
 export * as remotes from './remotes';
 export * as services from './services';
+export * as temporal from './temporal';
