+++ packages/sync-workflows/activities/do_process_sync_changes.ts
@@ -1,0 +1,14 @@
+import { SyncService } from 'sync-worker/services/sync_service';
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+export type DoProcessSyncChangesArgs = {};
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+export type DoProcessSyncChangesResult = {};
+
+export function createDoProcessSyncChanges(syncService: SyncService) {
+  return async function doProcessSyncChanges(args: DoProcessSyncChangesArgs): Promise<DoProcessSyncChangesResult> {
+    await syncService.processSyncChanges();
+    return {};
+  };
+}
