+++ packages/sync-workflows/workflows/index.ts
@@ -1,1 +1,2 @@
+export { processSyncChanges } from './process_sync_changes';
 export { runSync } from './run_sync';
